proyecto_abasolo/requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
bcrypt==4.1.3
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.4.3
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.1.7
django-bootstrap-v5==1.0.11
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cities==0.6.1
django-cors-headers==4.4.0
django-countries==7.5.1
django-countries-plus==2.2.0
django-crispy-forms==2.0
django-filter==23.5
django-fontawesome-5==1.0.18
django-model-utils==4.4.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
et-xmlfile==1.1.0
filelock==3.13.3
Flask==3.0.3
holidays==0.53
html5lib==1.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
lxml==5.3.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.2
paramiko==3.4.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pypdf==5.0.1
PyPDF2==3.0.1
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytweening==1.2.0
pytz==2023.3
pywhatkit==5.4
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==2.0.0
runserver==1.0.4
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
swapper==1.3.0
tinycss2==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wikipedia==1.4.0
xhtml2pdf==0.2.16
